+++ blocks/forms/forms.css
@@ -223,6 +223,7 @@ label.privacy {
 
 .hs-submit .actions button {
   width: 100%;
+  margin-bottom: 0;
 }
 
 /* hubspot form */
@@ -242,7 +243,7 @@ label.privacy {
 
 @media only screen and (min-width: 992px) {
   .hs-submit .actions {
-    margin: 10px auto 20px;
+    margin: 10px auto;
     max-width: 543px;
   }
 }
